ipalib: move json formatter to a separate file
To prevent cyclic imports, move JSON handling code to a separate file.
